Prevent wasted recalculations of TopBar
In regards to scalajs-react usage, I forsee two styles of use becoming
problematic in future and would like to take this opportunity to
demonstrate an appealing alternative. The problems:

1. React's shouldComponentRender hooks are used to determine whether a
component tree need be rerendered, analysed and compared with previous
vdom. Scastie currently does not make use of this feature and so when
any change occurs anywhere on the screen, the entire page is
re-calculated.

2. Unit tests currently don't exist but are in the backlog. Most
components currently ask for much more data that they require, and in a
singular, specific form. This will make unit testing hard.

By applying best practices of minimising dependencies and using
generic forms, it becomes trivial to address both issues described
above.

Using `TopBar` as an example, this commit demonstrates:

* how to reduce component props to a minimal subset
* how Reusability can be used to avoid wasted re-calculation
* how StateSnapshot can be used to abstract away state ownership
* how to use `.when` on vdom for concise code and improved readability
  (an aside)

It also replaces `clazz` with `cls` for the dom `class` attribute, which
avoids the need for special handling in the import.